With this commit, the code passes test 2. I solved a problem in OnSubflowRecv ; it was trying to recover both the data and the mapping from MpTcpSubflow:RecvWithMapping but the function would not map the DSN so it would always stay 0 and thus the meta RxBuffer NextRxSequence would never advance, leaving the connection stuck.
There is still the problem that both hosts send many duplicate acks: in
fact it sends a normal ack and then a normal ack + dataACk (DSS). I
believe the normal ack is sent from a call to a TcpSocketBase member.

Maybe it would also be best to inherit from TcpSocket rather than
TcpSocketBase.