before this commit, I had the problem where apparently everything was being sent, but once enabling DropTailQueue, I realized packets were dropped (look for "Queue full (at max packets) -- droppping pkt").
The retransmission timer would not work :/

acks were also ignored because the head of the buffr counts as the
acknowledged part and it was discared too quickly in NewAck via these
buggy lines:

if(m_nextTxSequence > mapping.TailSSN()) {

      m_txBuffer.DiscardUpTo(m_nextTxSequence);
}

I've started implementing a mechanism to discard more carefully the Tx
buffers.

doing so I noticed the buffers would need to undergo even more changes
if we want to simulate NR sack later on.

Be careful, simulation runs forever.