For the first time; I see data passing on the 2nd subflow.
A first problem is that the MP_JOIN + SYN/ACK is followed by a SYN/ACK.
Afterwards, I see in wireshark 2 packets without any Tcp header which is due to the
fact the TCP packet is bigger than the MSS and thus is fragmented then
reassembled.

TODO: cap the size of the packet with the MSS. Look at TcpSocketBase
for this.

Then the test crash because apparently some mappings overlap in Rx.