When looking at tcp-socket.h, one can see default values for: - SegmentSize (set to 536 but why ?) - InitialCwnd (set to 1)
The old mptcpSocketBase would set m_initialCwnd to 10.
As InitializeCwnd initializes the m_cWnd to m_initialCwnd *
m_segmentSize, it was 5360.
I've now left m_initialCwnd to 1 so the m_cWnd should be now init with 536.
I've modified subflow members like GetInitialCwnd so that it gets its
values from the meta.